Possible fix for rare crash on hitting enemy
Fixed a rare crash on shooting an enemy

Detailed issue description:
From what I've been told, the message system may call hooked functions on the frame(s) following the original Message event call. The message hooks in ACH were not designed with this behavior in mind, and were written with the assumption that the call would occur on the same frame as the hook. If the unit ceases to exist in the time (likely a single frame) between the time of the original Message call and the hooked function, the game can crash with an access violation due to operating on a NULL (removed and destroyed) Unit. (NULL Units are still considered non-nil values by Lua and must be checked with alive() )
In the reported instance of this crash (https://pastebin.com/mq15uxhC), the user was shooting at an enemy in a game of Holdout when the end screen was about to show up.