v31 addon mod xml/auto-update support,
Added an option to enable debug logs in the Misc menu.
Fixed a typo in the English localization where the "Reset Hitmarker Settings" button description would erroneously indicate that it was resetting the crosshair settings. (This is purely a localization issue and did not affect which settings would actually be reset.)
Behind the scenes optimization
Reduced divider size in the Misc menu