v31 cont
Oh yeah and enabled xml support for addons. Kind of major, that part.
Addons for ACH (installed in the usual place) can use a "mod.xml" file with the same specifications as a BeardLib "main.xml" file to easily implement auto-updates.

Cleaned code for whitespace, comments, and var names for internal consistency